Remove feedback link
The feedback link on geojson.io reports to my inbox, and I no longer
maintain / have commit access / provide support for geojson.io.

Feel free to replace it with a link to a Mapbox-operated feedback
mechanism, or anything else, but it's unfortunate for it to be
essentially `/dev/null` right now.